requirements.txt
fastapi
uvicorn
pydantic
requests
pillow
pytesseract
weaviate-client
httpx
openai
tqdm
python-dotenv
gunicorn